Refactor build automation
This commit refactors the build automation that has gone
through a few iterations over the years. This refactoring is
largely driven by the need to support Terraform's new
dependency locking and move away from hard locking
provider dependencies in the child modules to give users more
control over locked versions.

For this, the .terraform.lock.hcl has to be included in the
starter artifacts. Which requires running a terraform init,
inside the various root modules generated by dist.py,
before the module is compressed into the zip artifact.

Further, for historical reasons build automation was implemented
in 3 different tools and 2 different Dockerfiles. Initially
Google Cloud Build was used. Further parts where moved into GitHub
Actions to make the pipelines publicly visible.

To support local development, the Cloud Build and Actions were
wrapped into a Makefile.

Last but not least, when support for vendoring providers in the
image was added, this was implemented in a way that created a
dependency between the dist and build steps, because build now
required the output from dist to run terraform providers mirror.

To solve this, this commit moves the dist step, from the separate
GitHub action into a helper stage in the main Dockerfile. This ensures
terraform init and providers mirror are run with the correct lock file
and Terarform version to match the final container. This also has
the benefit, that now, running dist and build can be done with a
single docker build command.

Some testing (validate and plan) was implemented in the GitHub Actions
workflow. But integration tests were still in Google Cloud Build.

Going forward, the various test steps are targets in the Makefile. So
make can be run locally or from the pipeline to unify how tests are run
between local and automation.

After this commit GitHub Actions is the only CI/CD system used.
