--- .github/actions/build_artifacts/action.yaml
@@ -1,5 +1,0 @@
-name: "Build Artifacts"
-description: "Build the artifacts for the quickstart."
-runs:
-  using: 'docker'
-  image: 'Dockerfile'
