+++ Makefile
@@ -1,23 +1,86 @@
 _all: dist build
 
-GITHUB_REF := $(shell echo "refs/heads/"`git rev-parse --abbrev-ref HEAD`)
-GITHUB_SHA := $(shell echo `git rev-parse --verify HEAD^{commit}`)
+GIT_REF := $(shell echo "refs/heads/"`git rev-parse --abbrev-ref HEAD`)
+GIT_SHA := $(shell echo `git rev-parse --verify HEAD^{commit}`)
 
 dist:
-	docker build -t actions_build_artifacts:$(GITHUB_SHA) .github/actions/build_artifacts/
-	docker run --rm -e GITHUB_REF=$(GITHUB_REF) -e GITHUB_SHA=$(GITHUB_SHA) -v `pwd`/quickstart:/opt/quickstart:z actions_build_artifacts:$(GITHUB_SHA)
+	rm -rf quickstart/_dist
 
-build: dist
-	docker build --file oci/Dockerfile --progress plain -t kubestack/framework:local-$(GITHUB_SHA) . 
+	docker build \
+		--build-arg GIT_REF=${GIT_REF} \
+		--build-arg GIT_SHA=${GIT_SHA} \
+		--file oci/Dockerfile \
+		--progress plain \
+		-t dist-helper:latest \
+		--target dist-helper \
+		.
 
-test:
-	cloud-build-local --substitutions=_KBST_AUTH_AWS=${KBST_AUTH_AWS},_KBST_AUTH_GCLOUD=${KBST_AUTH_GCLOUD},_KBST_AUTH_AZ=${KBST_AUTH_AZ},_GITHUB_REF=$(GITHUB_REF),_GITHUB_SHA=$(GITHUB_SHA) --dryrun=false --config cloudbuild-test.yaml .
+	docker run \
+		--detach \
+		--name dist-helper \
+		--rm dist-helper:latest \
+		sleep 600
 
-test-cleanup:
-	cloud-build-local --substitutions=_KBST_AUTH_AWS=${KBST_AUTH_AWS},_KBST_AUTH_GCLOUD=${KBST_AUTH_GCLOUD},_KBST_AUTH_AZ=${KBST_AUTH_AZ},_GITHUB_REF=$(GITHUB_REF),_GITHUB_SHA=$(GITHUB_SHA) --dryrun=false --config cloudbuild-cleanup.yaml .
+	docker cp dist-helper:/quickstart/_dist quickstart/_dist
+	docker stop dist-helper
+	
 
-providers:
-	cloud-build-local --substitutions=_KBST_AUTH_AWS=${KBST_AUTH_AWS},_KBST_AUTH_GCLOUD=${KBST_AUTH_GCLOUD},_KBST_AUTH_AZ=${KBST_AUTH_AZ},_GITHUB_REF=$(GITHUB_REF),_GITHUB_SHA=$(GITHUB_SHA) --dryrun=false --config cloudbuild-providers.yaml .
+build:
+	docker build \
+		--build-arg GIT_REF=${GIT_REF} \
+		--build-arg GIT_SHA=${GIT_SHA} \
+		--file oci/Dockerfile \
+		--progress plain \
+		-t kubestack/framework:local-$(GIT_SHA) \
+		. 
 
-shell: build
-	docker run -ti --rm -e KBST_AUTH_AWS -e KBST_AUTH_AZ -e KBST_AUTH_GCLOUD -e HOME=/infra/tests/.user -v `pwd`:/infra:z --workdir /infra/tests kubestack/framework:local-$(GITHUB_SHA) bash
+validate: .init
+	docker exec \
+		test-container-$(GIT_SHA) \
+		entrypoint terraform validate
+
+test: validate
+	docker exec \
+		test-container-$(GIT_SHA) \
+		entrypoint terraform apply --input=false --auto-approve
+
+cleanup: .init
+	docker exec \
+		-ti \
+		test-container-$(GIT_SHA) \
+		entrypoint terraform destroy --input=false --auto-approve
+	$(MAKE) .stop-container
+
+shell: .init
+	docker exec \
+		-ti \
+		test-container-$(GIT_SHA) \
+		entrypoint bash
+
+.check-container:
+	docker inspect test-container-${GIT_SHA} > /dev/null || $(MAKE) .run-container
+
+.run-container: build
+	docker run \
+		--detach \
+		--name test-container-${GIT_SHA} \
+		--rm \
+		-v `pwd`:/infra:z \
+		-e KBST_AUTH_AWS \
+		-e KBST_AUTH_AZ \
+		-e KBST_AUTH_GCLOUD \
+		-e HOME=/infra/tests/.user \
+		--workdir /infra/tests \
+		kubestack/framework:local-$(GIT_SHA) \
+		sleep infinity
+
+.stop-container:
+	docker stop test-container-${GIT_SHA} || true
+
+.init: .check-container
+	docker exec \
+		test-container-$(GIT_SHA) \
+		entrypoint terraform init
+	docker exec \
+		test-container-$(GIT_SHA) \
+		entrypoint terraform workspace select ops
