--- cloudbuild-graph.yaml
@@ -1,62 +1,0 @@
-substitutions:
-    _HOME: /workspace/tests/.user
-    _TF_IN_AUTOMATION: "1"
-    _KBST_AUTH_AWS: ""
-    _KBST_AUTH_AZ: ""
-    _KBST_AUTH_GCLOUD: ""
-
-steps:
-- id: build dist action
-  name: 'gcr.io/cloud-builders/docker'
-  args:
-  - build
-  - -t
-  - actions_build_artifacts
-  - .github/actions/build_artifacts/
-
-- id: run dist action
-  name: 'actions_build_artifacts'
-  env:
-  - GITHUB_REF=$_GITHUB_REF
-  - GITHUB_SHA=$_GITHUB_SHA
-
-- id: docker build
-  name: 'gcr.io/cloud-builders/docker'
-  args:
-  - build
-  - -t
-  - kbst-infra-automation:bootstrap
-  - --file
-  - tests/ci-cd/Dockerfile
-  - .
-
-- id: authenticate
-  name: 'kbst-infra-automation:bootstrap'
-  dir: tests
-  env:
-  - HOME=$_HOME
-  - TF_IN_AUTOMATION=$_TF_IN_AUTOMATION
-  - KBST_AUTH_AWS=$_KBST_AUTH_AWS
-  - KBST_AUTH_AZ=$_KBST_AUTH_AZ
-  - KBST_AUTH_GCLOUD=$_KBST_AUTH_GCLOUD
-
-- id: terraform init
-  name: 'kbst-infra-automation:bootstrap'
-  dir: tests
-  env:
-  - HOME=$_HOME
-  - TF_IN_AUTOMATION=$_TF_IN_AUTOMATION
-  args:
-  - terraform
-  - init
-  - --input=false
-
-- id: terraform graph
-  name: 'kbst-infra-automation:bootstrap'
-  dir: tests
-  env:
-  - HOME=$_HOME
-  - TF_IN_AUTOMATION=$_TF_IN_AUTOMATION
-  args:
-  - terraform
-  - graph
